由于Spring使用的Json转换库默认只支持java.util.Date类型，不支持java.sql.Date类型，而客户端的Json库谷歌的配置只支持默认类型 故客户端获取的时间信息将以String形式传递，从而用到JsonFormat注释。 暂时取消认证码的设定，而直接返回登录以后的User信息。

+++ src/main/java/otakuplus/straybird/othellogameserver/network/Logout.java
@@ -1,15 +1,13 @@
 package otakuplus.straybird.othellogameserver.network;
 
-import otakuplus.straybird.othellogameserver.network.Logout;
-
 public class Logout {
-	private int userId;
+	private Long userId;
 
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
